Remember admin's username and password using cookie service

+++ src/app/component/login/login.component.ts
@@ -2,8 +2,11 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder,Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { CookieService } from 'ngx-cookie-service';
+import { Admin } from 'src/app/model/admin';
 
 import { AdminService } from 'src/app/service/admin.service';
+import { AuthService } from 'src/app/service/auth.service';
 
 @Component({
   selector: 'app-login',
@@ -13,14 +16,19 @@ import { AdminService } from 'src/app/service/admin.service';
 export class LoginComponent implements OnInit {
 
    loginForm: FormGroup;
+   public admin = new Admin();
+   public isChecked ;
 
-  constructor( private formBuilder: FormBuilder ,private adminService: AdminService, private router : Router) 
-  { }
+  constructor( private formBuilder: FormBuilder ,private adminService: AdminService, private router : Router , private cookie : CookieService ) 
+  { 
+    
+  }
 
   ngOnInit(): void {
     this.loginForm = this.formBuilder.group({
       username:['' , Validators.required],
-      password:['', Validators.required]
+      password:['', Validators.required],
+      remember:['']
     })
   }
 
@@ -31,12 +39,25 @@ export class LoginComponent implements OnInit {
            return admin.username === this.loginForm.value.username && admin.password === this.loginForm.value.password
          });
          if (user){
-           alert('Login success');
-           localStorage.setItem('token','jkj');
-           this.router.navigate(['students']);
+             if (confirm('do you want to save your password?')){
+              
+              this.cookie.set("username",this.loginForm.value.username);
+              this.cookie.set("password",this.loginForm.value.password);
+              alert('Login success');
+              localStorage.setItem('token','jkj');
+              this.router.navigate(['students']);
+             }else{
+              alert('Login success');
+              localStorage.setItem('token','jkj');
+              this.router.navigate(['students']);
+             }
+            
+           }
+       
          
-         }else{
+         else{
           alert('username or password incorrect!');
+         
          }
        },
        err=>{
@@ -45,6 +66,8 @@ export class LoginComponent implements OnInit {
 
      )
   }
+
+ 
 
  
 }
